backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.8
django-colorfield==0.11.0
django-filter==23.5
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==6.1.0
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.4
